refactor(examples): clarify protocol boundary and low-level API rationale
Document why consent examples use the low-level Server API instead of
withMCPI: McpServer.registerTool strips unknown keys via zod schema
validation, which silently drops _mcpi_delegation before the handler
sees it. Delegation-protected tools require raw args.

- Add architecture note to server.ts headers (both consent examples)
- Add section markers to consent-full/server.ts for scannable structure
- Simplify identity loading using generateIdentity from middleware
- Add Architecture section to consent-full README with two-server diagram
- Reorder file structure table to highlight consent-server.ts as showcase
- Update test imports to use middleware/index.js for consistency
